src/main/java/vives/bancovives/security/userauthentication/AuthenticationServiceImpl.java
package vives.bancovives.security.userauthentication;

import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;
import vives.bancovives.rest.users.auth.AuthUsersService;
import vives.bancovives.rest.users.dto.input.UserRequest;
import vives.bancovives.rest.users.exceptions.UserBadRequest;
import vives.bancovives.rest.users.services.UsersService;
import vives.bancovives.security.jwt.JwtService;
import vives.bancovives.security.model.JwtAuthResponse;

/**
 * Implementación de nuestro servicio de autenticación
 */
@Service
@Slf4j
public class AuthenticationServiceImpl implements AuthenticationService {
    private final AuthUsersService userService;
    private final UsersService anotherUsersService;
    private final PasswordEncoder passwordEncoder;
    private final JwtService jwtService;
    private final AuthenticationManager authenticationManager;

    @Autowired
    public AuthenticationServiceImpl(AuthUsersService userService, UsersService anotherUsersService, PasswordEncoder passwordEncoder,
                                     JwtService jwtService, AuthenticationManager authenticationManager) {
        this.userService = userService;
        this.anotherUsersService = anotherUsersService;
        this.passwordEncoder = passwordEncoder;
        this.jwtService = jwtService;
        this.authenticationManager = authenticationManager;
    }

    /**
     * Registra un usuario
     *
     * @param request datos del usuario
     * @return Token de autenticación
     */
    @Override
    public JwtAuthResponse signUp(UserRequest request) {
        log.info("Creando usuario: {}", request);
        String encodedPassword = passwordEncoder.encode(request.getPassword());
        request.setPassword(encodedPassword);
        var userStored = anotherUsersService.save(request);
        return JwtAuthResponse.builder().token(jwtService.generateToken(userStored)).build();
    }

    /**
     * Autentica un usuario
     *
     * @param request datos del usuario
     * @return Token de autenticación
     */
    @Override
    public JwtAuthResponse signIn(UserRequest request) {
        log.info("Autenticando usuario: {}", request);
        // Autenticamos y devolvemos el token
        authenticationManager.authenticate(
                new UsernamePasswordAuthenticationToken(request.getUsername(), request.getPassword()));
        var user = userService.loadUserByUsername(request.getUsername());
        if (!passwordEncoder.matches(request.getPassword(), user.getPassword())) throw new UserBadRequest("La contraseña no es correcta");
        var jwt = jwtService.generateToken(user);
        return JwtAuthResponse.builder().token(jwt).build();
    }
}